feature: modularize the code for better separation of concerns. index.js is now an orchestrator
